Added partial lap support - User added partial laps - Auto calculated partial laps
https://github.com/daufderheide/racecoordinator_ai/issues/134
https://github.com/daufderheide/racecoordinator_ai/issues/133

Squashed commit of the following:

commit d5844a59730998952adc14eb1eb5e9b585dc43b1
Author: Dave <contact@racecoordinator.net>
Date:   Fri May 1 22:18:28 2026 -0400

    Added "do not allow finish auto segments" support

    https://github.com/daufderheide/racecoordinator_ai/issues/134
    https://github.com/daufderheide/racecoordinator_ai/issues/133

commit a137f405e5187250f41c72952dfa91f08125e707
Author: Dave <contact@racecoordinator.net>
Date:   Fri May 1 20:09:50 2026 -0400

    Added ability to add negative user laps with shift-left click on the lap column

commit a2cc741ed05541eb7453cd73e4deed5cfa0dcd26
Author: Dave <contact@racecoordinator.net>
Date:   Fri May 1 16:57:36 2026 -0400

    Fixed more lint issues

commit 6bb4e6147e48f695874f1d4e8e0168edbe762db6
Author: Dave <contact@racecoordinator.net>
Date:   Fri May 1 16:51:19 2026 -0400

    Added lint/checkstyle checks for unused imports/vaariables/etc
    Added checktyle check for fully qualified domaain names in variables
    Updated CI to hopefully run these checks
    Added section in the readme for lint checks
    Massive code cleanup

commit daebf8863eaa8613628bd5b3df7eac3bb861b918
Author: Dave <contact@racecoordinator.net>
Date:   Fri May 1 13:18:10 2026 -0400

    Fixed missing average lap time on the race day page
    Added tests to ensure average lap time calculation is correct
    Fixed console error on race day page (angular usage issue)
    Removed unused imports and variables from race day component

commit ee3a07bd3aac21c97c3bad13aa56762e120b669f
Author: Dave <contact@racecoordinator.net>
Date:   Fri May 1 11:15:07 2026 -0400

    Added new tests and updated screendiffs

commit de1ec38a13cd34ff35ab7e678f82a36604b05eab
Author: Dave <contact@racecoordinator.net>
Date:   Fri May 1 01:14:49 2026 -0400

    First pass at partial laps.  Left clicking on the lap column adds 0.25 laps to the drivers total

+++ server/src/main/java/com/antigravity/models/GlobalStatistics.java
@@ -23,7 +23,7 @@ public class GlobalStatistics {
 
   @BsonProperty("total_laps")
   @JsonProperty("total_laps")
-  private int totalLaps;
+  private double totalLaps;
 
   @BsonProperty("total_race_time_ms")
   @JsonProperty("total_race_time_ms")
@@ -145,7 +145,7 @@ public class GlobalStatistics {
       @BsonId @JsonProperty("_id") ObjectId id,
       @BsonProperty("race_entity_id") @JsonProperty("race_entity_id") String raceEntityId,
       @BsonProperty("total_races") @JsonProperty("total_races") int totalRaces,
-      @BsonProperty("total_laps") @JsonProperty("total_laps") int totalLaps,
+      @BsonProperty("total_laps") @JsonProperty("total_laps") double totalLaps,
       @BsonProperty("total_race_time_ms") @JsonProperty("total_race_time_ms") long totalRaceTimeMs,
       @BsonProperty("fastest_lap_time") @JsonProperty("fastest_lap_time") double fastestLapTime,
       @BsonProperty("fastest_lap_driver_name") @JsonProperty("fastest_lap_driver_name")
@@ -258,15 +258,15 @@ public class GlobalStatistics {
     this.totalRaces++;
   }
 
-  public int getTotalLaps() {
+  public double getTotalLaps() {
     return totalLaps;
   }
 
-  public void setTotalLaps(int totalLaps) {
+  public void setTotalLaps(double totalLaps) {
     this.totalLaps = totalLaps;
   }
 
-  public void addLaps(int laps) {
+  public void addLaps(double laps) {
     this.totalLaps += laps;
   }
